task-5/requirements.txt
langchain-chroma>=0.1.0
langchain-cohere>=0.1.0
python-dotenv>=1.0.0
